Add files via upload
We discussed the basics of Python in the previous lecture. We learned that Python
comes with various built in batteries or modules. These batteries or modules
perform various specialized operations. The modules can be used to perform
computation, database management, web server etc. Since this lecture is focused
on creating scientific applications, we limit our focus to Python modules that allow
computation such as scipy, numpy, matplotlib, pandas. We discuss the relevance of
each of these modules and explain their use with examples. We also discuss
creation of new Python modules.